+++ modules/api.js
@@ -1,0 +1,14 @@
+export async function importComments() {
+    return fetch('https://wedev-api.sky.pro/api/v1/anton-nikonov/comments', {
+        method: 'GET',
+    }).then((response) => {
+        return response.json()
+    })
+}
+
+export async function exportComments(data) {
+    fetch('https://wedev-api.sky.pro/api/v1/anton-nikonov/comments', {
+        method: 'POST',
+        body: JSON.stringify(data),
+    })
+}
